test(agentic-memory): isolate tests and fix OpenSearch 3.3 mocking
