+++ tests/tools/agentic_memory_data.py
@@ -1,5 +1,5 @@
 import pytest
-from tools.tool_params import MemoryType, PayloadType
+from tools.agentic_memory.params import MemoryType, PayloadType
 
 
 # --- Agentic Memory: parametrized test input cases for Pydantic model validation ---
@@ -525,4 +525,33 @@ CREATE_CONTAINER_HAPPY_PATH_CASES = [
         {'memory_container_id': 'MinimalContainer123', 'status': 'created'},
         id='minimal_config',
     ),
+]
+
+
+__all__ = [
+    'UPDATE_HAPPY_PATH_CASES',
+    'ADD_MEMORIES_HAPPY_PATH_CASES',
+    'CREATE_SESSION_HAPPY_PATH_CASES',
+    'GET_MEMORY_HAPPY_PATH_CASES',
+    'SEARCH_SESSIONS_RESPONSE',
+    'SEARCH_LONG_TERM_RESPONSE',
+    'SEARCH_WORKING_COMPLEX_RESPONSE',
+    'SEARCH_HISTORY_RESPONSE',
+    'SEARCH_NO_SORT_RESPONSE',
+    'SEARCH_SESSION_ID_RESPONSE',
+    'SEARCH_MEMORY_HAPPY_PATH_CASES',
+    'DELETE_MEMORY_ID_HAPPY_PATH_CASES',
+    'DELETE_QUERY_WORKING_RESP',
+    'DELETE_QUERY_LONG_TERM_RESP',
+    'DELETE_QUERY_SESSIONS_RESP',
+    'DELETE_QUERY_COMPLEX_RESP',
+    'DELETE_QUERY_NO_RESULTS_RESP',
+    'DELETE_MEMORY_QUERY_HAPPY_PATH_CASES',
+    'BASIC_CONFIG_PAYLOAD',
+    'BASIC_CONFIG_EXPECTED_BODY',
+    'ADVANCED_CONFIG_PAYLOAD',
+    'ADVANCED_CONFIG_EXPECTED_BODY',
+    'MINIMAL_CONFIG_PAYLOAD',
+    'MINIMAL_CONFIG_EXPECTED_BODY',
+    'CREATE_CONTAINER_HAPPY_PATH_CASES',
 ]
